+++ spec/lib/fediverse/request_spec.rb
@@ -20,5 +20,21 @@ module Fediverse
         end
       end
     end
+
+    describe '.dereference' do
+      context 'when value is a hash' do
+        it 'returns the hash' do
+          expect(described_class.dereference({})).to eq({})
+        end
+      end
+
+      context 'when value is a string' do
+        it 'returns a Hash' do
+          VCR.use_cassette 'fediverse/request/get_get_200' do
+            expect(described_class.dereference('https://mamot.fr/users/mtancoigne')).to be_a Hash
+          end
+        end
+      end
+    end
   end
 end